# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /community-posts:
    get:
      operationId: getCommunityPosts
      tags: [community]
      summary: Get approved active community posts
      parameters:
        - name: cityId
          in: query
          required: true
          schema:
            type: string
        - name: neighborhood
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/CommunityPostType"
      responses:
        "200":
          description: Active approved posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPostList"
        "400":
          description: Invalid city or post type
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Submit a community post for review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityPostInput"
      responses:
        "201":
          description: Submitted post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPostSubmission"
        "400":
          description: Invalid community post
        "401":
          description: Authentication is required
        "403":
          description: Registration is required to post
  /community-posts/moderation:
    get:
      operationId: getCommunityModerationPosts
      tags: [community]
      summary: Get the community moderation queue
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, all]
            default: pending
      responses:
        "200":
          description: Moderation queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPostList"
        "401":
          description: Authentication is required
        "403":
          description: An editor role is required
  /community-posts/moderation/{id}:
    patch:
      operationId: decideCommunityPost
      tags: [community]
      summary: Approve or reject a community post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityPostDecision"
      responses:
        "200":
          description: Updated moderation item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
        "400":
          description: Invalid moderation decision
        "401":
          description: Authentication is required
        "403":
          description: An editor role is required
        "404":
          description: Post not found
  /community-posts/{id}/participation:
    put:
      operationId: toggleCommunityPostParticipation
      tags: [community]
      summary: Add or remove the signed-in resident's interest or attendance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityPostParticipationInput"
      responses:
        "200":
          description: Updated privacy-safe participation summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPostParticipationResponse"
        "400":
          description: Invalid participation action or attendance requested for a non-event post
        "401":
          description: Authentication is required
        "403":
          description: Registration is required to participate
        "404":
          description: Approved active post not found
...
components:
...
  schemas:
...
    CommunityPostType:
      type: string
      enum: [event, question, help_offer, help_request, tip, announcement]
    CommunityPostStatus:
      type: string
      enum: [pending, approved, rejected]
    CommunityPost:
      type: object
      properties:
        id:
          type: number
        cityId:
          type: string
        neighborhood:
          type: ["string", "null"]
        type:
          $ref: "#/components/schemas/CommunityPostType"
        title:
          type: string
        body:
          type: string
        startsAt:
          type: ["string", "null"]
        expiresAt:
          type: string
        status:
          $ref: "#/components/schemas/CommunityPostStatus"
        reviewNote:
          type: ["string", "null"]
        createdAt:
          type: string
        interestCount:
          type: number
        attendanceCount:
          type: number
        interestedByMe:
          type: boolean
        attendingByMe:
          type: boolean
      required: [id, cityId, type, title, body, expiresAt, status, createdAt, interestCount, attendanceCount, interestedByMe, attendingByMe]
    CommunityPostInput:
      type: object
      properties:
        cityId:
          type: string
          minLength: 1
        neighborhood:
          type: ["string", "null"]
        type:
          $ref: "#/components/schemas/CommunityPostType"
        title:
          type: string
          minLength: 3
          maxLength: 120
        body:
          type: string
          minLength: 10
          maxLength: 2000
        startsAt:
          type: ["string", "null"]
        expiresAt:
          type: string
      required: [cityId, type, title, body, expiresAt]
    CommunityPostSubmission:
      type: object
      properties:
        id:
          type: number
        status:
          $ref: "#/components/schemas/CommunityPostStatus"
        message:
          type: string
      required: [id, status, message]
    CommunityPostDecision:
      type: object
      properties:
        decision:
          type: string
          enum: [approve, reject]
        reviewNote:
          type: string
          maxLength: 500
      required: [decision]
    CommunityPostList:
      type: array
      items:
        $ref: "#/components/schemas/CommunityPost"
    CommunityPostParticipationAction:
      type: string
      enum: [interested, attending]
    CommunityPostParticipationInput:
      type: object
      properties:
        action:
          $ref: "#/components/schemas/CommunityPostParticipationAction"
        active:
          type: boolean
      required: [action, active]
    CommunityPostParticipationResponse:
      type: object
      properties:
        postId:
          type: number
        action:
          $ref: "#/components/schemas/CommunityPostParticipationAction"
        active:
          type: boolean
        interestCount:
          type: number
        attendanceCount:
          type: number
        interestedByMe:
          type: boolean
        attendingByMe:
          type: boolean
      required: [postId, action, active, interestCount, attendanceCount, interestedByMe, attendingByMe]
...